Split translator into several files.
