+++ src/mmt-mizar/src/info/kwarc/mmt/mizar/newxml/translator/Justification.scala
@@ -1,0 +1,7 @@
+package info.kwarc.mmt.mizar.newxml.translator
+
+import info.kwarc.mmt.mizar.newxml.syntax.Justification
+
+object justificationTranslator {
+  def translate_Justification(just:Justification) = { ??? }
+}
